# open files before `read`, `modify`, `collect` and `Array`