--- a/src/array.jl
+++ b/src/array.jl
@@ -28,7 +28,7 @@ missingval(A::AbstractRaster) = A.missingval
 filename(A::AbstractRaster) = filename(parent(A))
 filename(A::AbstractArray) = nothing # Fallback
 
-cleanreturn(A::AbstractRaster) = modify(cleanreturn, A)
+cleanreturn(A::AbstractRaster) = rebuild(A, cleanreturn(parent(A)))
 cleanreturn(x) = x
 
 isdisk(A::AbstractRaster) = parent(A) isa DiskArrays.AbstractDiskArray
@@ -134,6 +134,9 @@ end
 # Base methods
 
 Base.parent(A::AbstractRaster) = A.data
+# Make sure a disk-based Raster is open before Array/collect
+Base.Array(A::AbstractRaster) = open(O -> Array(parent(O)), A)
+Base.collect(A::AbstractRaster) = open(O -> collect(parent(O)), A)
 
 """
     open(f, A::AbstractRaster; write=false)