--- a/src/array.jl
+++ b/src/array.jl
@@ -108,6 +108,14 @@ function DD.rebuild(A::AbstractRaster;
     rebuild(A, data, dims, refdims, name, metadata, missingval)
 end
 
+function DD.modify(f, A::AbstractRaster)
+    newdata = open(A) do O
+        f(parent(O))
+    end
+    size(newdata) == size(A) || error("$f returns an array with size $(size(newdata)) when the original size was $(size(A))")
+    return rebuild(A, newdata)
+end
+
 function DD.DimTable(As::Tuple{<:AbstractRaster,Vararg{<:AbstractRaster}}...)
     DimTable(DimStack(map(read, As...)))
 end