Fixes the performance of `read`, `modify`, `collect` and `Array` by using `open` on the file first. This should have always been the implementation.

@kongdd this PR is a huge improvement to load time without needing the `lazy` flag. We might still need it, as there is always some memory overhead to using lazy DiskArrays methods. But this is a start.

Also notice that in your timings in #254 you were including compilation time in the first run. The `Raster` object is a little complicated and takes a few seconds to compile, and GDAL/ArchGDAL also compile in that first run. We can try to fix that too at some stage, but its a separate issue to this one.

For now, the second and following runs are the real load time without compilation.

This PR:
```julia
julia> using Rasters

julia> f = "/home/raf/Downloads/PMLV2_yearly_G010_v014_2017-01-01.tif"
"/home/raf/Downloads/PMLV2_yearly_G010_v014_2017-01-01.tif"

julia> @time r = read(Raster(f)) # First run: compilation + load time
  4.375702 seconds (10.60 M allocations: 1.133 GiB, 9.57% gc time, 74.30% compilation time)

julia> @time r = read(Raster(f)); # Second run - the actual load time
  0.995884 seconds (2.30 k allocations: 576.877 MiB, 0.72% gc time)
```

Master:
```julia

julia> using Rasters

julia> f = "/home/raf/Downloads/PMLV2_yearly_G010_v014_2017-01-01.tif"
"/home/raf/Downloads/PMLV2_yearly_G010_v014_2017-01-01.tif"

julia> @time r = read(Raster(f));
  8.612977 seconds (9.55 M allocations: 1.072 GiB, 2.60% gc time, 32.18% compilation time)

julia> @time r = read(Raster(f));
  5.713882 seconds (1.01 k allocations: 576.819 MiB, 2.00% gc time)
```