@kongdd 2022-03-29
Amazing. Small change, big improvement. 
But, actually, I can't figure out why current script improve the performance.