@rafaqz 2022-03-29
@kongdd to make this even more confusing for you ;)

The PR now alters `modify` so it happens at a lower level. `modify` is the DimensionalData.jl way of swapping the internal array of an object for something else, a bit like `set` but just for the array data. Its another way of making it easier to deal with these imuutable objects.